Set IWebHostBuilder.ContentRoot
The default value for the `ContentRoot` is the project's root directory.
When the binaries are copied to the `test-runner` image, the souce code
directory is omitted.  Then, when the web application attempts to start
and read from the directory that is not present, and exception is
thrown.

To avoid this exception, the `CustomWebApplicationFactory` sets the
`ContentRoot` to the test assembly's directory.